Test gradient package works locally.
